Added a variation model for infinite-alleles mutation within the unit interval (0,1) and a Wright-Fisher drift model implementing it.  Given the ITraitFactory instance, everything else just works, and it gives Doubles for trait ID's, and the K_n statistic is correct.

+++ src/main/java/org/madsenlab/sim/tf/traits/RealTraitProvider.java
@@ -26,8 +26,8 @@ public class RealTraitProvider implements Provider<ITrait> {
     @Inject
     private ISimulationModel model;
 
-    public ITrait<Float> get() {
-        ITrait<Float> trait = new UnstructuredTrait<Float>();
+    public ITrait<Double> get() {
+        ITrait<Double> trait = new UnstructuredTrait<Double>();
         trait.setSimulationModel(model);
         return trait;
     }
